+++ main.py
@@ -5,7 +5,9 @@ from fastapi import FastAPI
 from location_services import find_coordinate, Address, Coordinate
 from hsl_services import get_routes
 
-app = FastAPI()
+from routers import hsl_v1
+
+app = FastAPI(title="HSL Timer API")
 
 @app.get("/")
 def read_root():
@@ -13,28 +15,4 @@ def read_root():
         'message': "Welcome to HSL Timer API"
     }
 
-@app.get("/coordinateFor")
-async def get_coordinate(street: str, city: str, postalCode: str):
-    address = Address(street, city, 'Finland', postalCode)
-    coordinate = await find_coordinate(address)
-    return {
-        'lat': coordinate.lat,
-        'lon': coordinate.lon
-    }
-
-@app.get("/routesForCoordinate")
-async def get_routes_for_coordinate(lat: float, lon: float):
-    coordinate = Coordinate(lat, lon)
-    routes = await get_routes(coordinate)
-    return {
-        'routes': sorted(routes, key=lambda r: r.arrive_at)
-    }
-
-@app.get("/routesForAddress")
-async def get_routes_for_address(street: str, city: str, postalCode: str):
-    address = Address(street, city, 'Finland', postalCode)
-    coordinate = await find_coordinate(address)
-    routes = await get_routes(coordinate)
-    return {
-        'routes': sorted(routes, key=lambda r: r.arrive_at)
-    }
+app.include_router(hsl_v1.router, prefix='/v1', tags=['hsl', 'v1'])
